Preparing for  New Graphic Standards
