--- resources/Ladybug_Body_Template.py
@@ -1,61 +1,0 @@
-
-def checkTheInputs():
-	#....(INSERT INPUT CHECKING FUNCTIONS HERE)....
-
-	return False
-
-
-def main():
-	#....(INSERT MAIN COMPONENTS FUNCTIONS HERE)....
-
-	return -1
-
-
-
-
-#If Honeybee or Ladybug is not flying or is an older version, give a warning.
-initCheck = True
-
-#Ladybug check.
-if not sc.sticky.has_key('ladybug_release') == True:
-    initCheck = False
-    print "You should first let Ladybug fly..."
-    ghenv.Component.AddRuntimeMessage(w, "You should first let Ladybug fly...")
-else:
-    try:
-        if not sc.sticky['ladybug_release'].isCompatible(ghenv.Component): pass
-    except:
-        initCheck = False
-        warning = "You need a newer version of Ladybug to use this compoent." + \
-        "Use updateLadybug component to update userObjects.\n" + \
-        "If you have already updated userObjects drag Ladybug_Ladybug component " + \
-        "into canvas and try again."
-        w = gh.GH_RuntimeMessageLevel.Warning
-        ghenv.Component.AddRuntimeMessage(w, warning)
-
-
-#Honeybee check.
-if not sc.sticky.has_key('honeybee_release') == True:
-	initCheck = False
-    print "You should first let Honeybee fly..."
-    ghenv.Component.AddRuntimeMessage(w, "You should first let Honeybee fly...")
-else:
-	try:
-        if not sc.sticky['honeybee_release'].isCompatible(ghenv.Component): pass
-    except:
-		initCheck = False
-        warning = "You need a newer version of Honeybee to use this compoent." + \
-        "Use updateHoneybee component to update userObjects.\n" + \
-        "If you have already updated userObjects drag Honeybee_Honeybee component " + \
-        "into canvas and try again."
-        w = gh.GH_RuntimeMessageLevel.Warning
-        ghenv.Component.AddRuntimeMessage(w, warning)
-
-
-#If the intital check is good, run the component.
-if initCheck:
-	checkData = checkTheInputs():
-	if checkData:
-		result = main()
-		if result != -1:
-			output = result
