--- resources/Ladybug_Header_Template.py
@@ -1,50 +1,0 @@
-# Ladybug: A Plugin for Environmental Analysis (GPL) started by Mostapha Sadeghipour Roudsari
-# 
-# This file is part of Ladybug.
-# 
-# Copyright (c) 2013-2015, ....(YOUR NAME).... <....(YOUR EMAIL)....> 
-# Ladybug is free software; you can redistribute it and/or modify 
-# it under the terms of the GNU General Public License as published 
-# by the Free Software Foundation; either version 3 of the License, 
-# or (at your option) any later version. 
-# 
-# Ladybug is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of 
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
-# GNU General Public License for more details.
-# 
-# You should have received a copy of the GNU General Public License
-# along with Ladybug; If not, see <http://www.gnu.org/licenses/>.
-# 
-# @license GPL-3.0+ <http://spdx.org/licenses/GPL-3.0+>
-
-
-"""
-Use this component to ....(COMPONENT DESCRIPTION)....
-_
-....(REFERENCES + PAPERS)....
--
-Provided by Ladybug 0.0.61
-    
-    Args:
-        _input1: A ....(DATE TYPE).... that represents ....(INPUT DESCRIPTION).... .
-        _input2_: A ....(DATE TYPE).... that represents ....(INPUT DESCRIPTION).... .  If nothing is connected here, a default of ....(DEFAULT VALUE).... will be used.
-        input3_: The output from the ....(OTHER LB COMPONENT).... component.  Use this to ....(INPUT DESCRIPTION).... .
-    Returns:
-        readMe!: ...
-        output1: A ....(DATE TYPE).... that represents ....(OUTPUT DESCRIPTION).... .
-		output2: A ....(DATE TYPE).... that represents ....(OUTPUT DESCRIPTION).... .
-		output3: A ....(DATE TYPE).... that represents ....(OUTPUT DESCRIPTION).... .
-"""
-
-ghenv.Component.Name = "Ladybug_....(COMPONENT NAME)...."
-ghenv.Component.NickName = '....(COMPONENT NICKNAME)....'
-ghenv.Component.Message = 'VER 0.0.61\nNOV_05_2015' #Change this date to be that of your commit or pull request.
-ghenv.Component.Category = "Ladybug"
-ghenv.Component.SubCategory = "6 | WIP"
-#Change the following date to be that of the LB version during your commit or pull request:
-#compatibleLBVersion = VER 0.0.59\nFEB_01_2015
-#Change the following date to be that of the HB version of your commit or pull request (or get rid of the follwoing if your component is a part of Ladybug):
-#compatibleHBVersion = VER 0.0.56\nFEB_01_2015
-try: ghenv.Component.AdditionalHelpFromDocStrings = "1"
-except: pass
